Complete refactoring of portfolio and portfolio details views
