+++ ftt/ui/portfolio/signals.py
@@ -1,0 +1,7 @@
+from PySide6.QtCore import QObject, Signal
+
+
+class PortfolioSignals(QObject):
+    portfolioChanged = Signal(int)
+    portfolioVersionSelected = Signal(int)
+    portfolioVersionInDBUpdated = Signal()
